+++ src/app/components/TopBar/container.js
@@ -2,7 +2,7 @@ import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
 
 import TopBar from "./TopBar";
-import { toggleDrawer } from "../../modules/SideDrawer/actions";
+import { toggleDrawer } from "../SideDrawer/actions";
 import { handleCalendarChange } from "../../modules/DailyForm/actions";
 
 const mapStateToProps = state => ({
